updated capabilityinfo tests
